Update phoenix
Note: weird telemetry error when running tests but app runs fine and all tests still pass. Rolling forward and backward over this phoenix change does not seem to resolve the issue. The error mentions the NR agent so if it continues to be an issue, rolling the agent forward or back is a good next step.